screen-orientation/lock-basic.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  screen.orientation.unlock();
}, "Test that screen.orientation.unlock() doesn't throw when there is no lock");

test(() => {
  const value = screen.orientation.unlock();
  assert_equals(value, undefined);
}, "Test that screen.orientation.unlock() returns a void value");

promise_test(t => {
  return screen.orientation.lock('any').then(value => {
    assert_equals(value, undefined);
  });
}, "Test that screen.orientation.lock returns a promise which will be fulfilled with a void value.");

promise_test(t => {
  const orientations = [
    'any',
    'portrait',
    'landscape',
    'portrait-secondary',
    'landscape-primary',
    'landscape-secondary',
    'portrait-primary',
  ];
  const promises = orientations.map(type => {
    const returnType = screen.orientation.lock(type);
    assert_true(returnType instanceof Promise);
  });
  return Promise.all(promises).finally(() => {
    screen.orientation.unlock();
  });
}, "Test that screen.orientation.lock returns a pending promise.");

test(() => {
  const previousOrientation = screen.orientation;
  assert_equals(screen.orientation, previousOrientation);
}, "Test that screen.orientation.lock() is actually async");
</script>
